Move ReactDOMLegacy implementation into RootFB
Only the FB entry point has legacy mode now so we can move the remaining
code in there.
