+++ packages/react-devtools-shell/src/e2e-regression/app-legacy.js
@@ -15,6 +15,7 @@ function mountApp(App: () => React$Node) {
 
   ((document.body: any): HTMLBodyElement).appendChild(container);
 
+  // $FlowFixMe[prop-missing]: These are removed in 19.
   ReactDOM.render(<App />, container);
 }
 function mountTestApp() {
