+++ packages/react-dom/index.js
@@ -15,11 +15,8 @@ export {
   createRoot,
   hydrateRoot,
   flushSync,
-  render,
-  unmountComponentAtNode,
   unstable_batchedUpdates,
   unstable_createEventHandle,
-  unstable_renderSubtreeIntoContainer,
   unstable_runWithPriority, // DO NOT USE: Temporarily exposed to migrate off of Scheduler.runWithPriority.
   useFormStatus,
   useFormState,
