+++ packages/react-dom/index.stable.js
@@ -13,10 +13,7 @@ export {
   createRoot,
   hydrateRoot,
   flushSync,
-  render,
-  unmountComponentAtNode,
   unstable_batchedUpdates,
-  unstable_renderSubtreeIntoContainer,
   useFormStatus,
   useFormState,
   prefetchDNS,
