+++ packages/react-dom/unstable_testing.js
@@ -10,11 +10,8 @@
 export {
   createPortal,
   flushSync,
-  render,
-  unmountComponentAtNode,
   unstable_batchedUpdates,
   unstable_createEventHandle,
-  unstable_renderSubtreeIntoContainer,
   unstable_runWithPriority, // DO NOT USE: Temporarily exposed to migrate off of Scheduler.runWithPriority.
   useFormStatus,
   useFormState,
