+++ packages/react/src/__tests__/ReactJSXRuntime-test.js
@@ -10,12 +10,10 @@
 'use strict';
 
 let React;
-let ReactDOM;
 let ReactDOMClient;
 let JSXRuntime;
 let JSXDEVRuntime;
 let act;
-let findDOMNode;
 
 // NOTE: Prefer to call the JSXRuntime directly in these tests so we can be
 // certain that we are testing the runtime behavior, as opposed to the Babel
@@ -27,11 +25,8 @@ describe('ReactJSXRuntime', () => {
     React = require('react');
     JSXRuntime = require('react/jsx-runtime');
     JSXDEVRuntime = require('react/jsx-dev-runtime');
-    ReactDOM = require('react-dom');
     ReactDOMClient = require('react-dom/client');
     act = require('internal-test-utils').act;
-    findDOMNode =
-      ReactDOM.__SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED.findDOMNode;
   });
 
   it('allows static methods to be called using the type property', () => {
@@ -133,9 +128,9 @@ describe('ReactJSXRuntime', () => {
 
     const outer = container.firstChild;
     if (__DEV__) {
-      expect(findDOMNode(outer).className).toBe('moo');
+      expect(outer.className).toBe('moo');
     } else {
-      expect(findDOMNode(outer).className).toBe('quack');
+      expect(outer.className).toBe('quack');
     }
   });
